requirements.txt
streamlit==1.37.1
psycopg2-binary==2.9.9
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
